clean ds_widget module
this commit moves all dialogs into the dialogs module, and all plotmethod widgets into the plotmethods module
